backend/requirements.txt
# Core Backend Framework
fastapi>=0.110.0,<1.0        # Web framework for building APIs
uvicorn[standard]>=0.29.0    # ASGI server to run FastAPI apps


# Database (PostgreSQL)
sqlalchemy>=2.0.25           # ORM for interacting with PostgreSQL
psycopg2-binary>=2.9.9       # PostgreSQL database driver
alembic>=1.13.1              # Database migrations (schema versioning)


# Data Validation & Settings
pydantic>=2.6.0              # Data validation & serialization
pydantic-settings>=2.2.0     # Environment-based config management


# Authentication & Security
python-jose[cryptography]>=3.3.0  # JWT token handling
passlib[bcrypt]>=1.7.4            # Password hashing
bcrypt==3.2.2                   # Secure hashing backend


# Forms & Email Validation
python-multipart>=0.0.9       # Handle form-data (file uploads, login forms)
email-validator>=2.1.1        # Validate email addresses
